Created a main menu and function to display all plants

+++ Plant.cs
@@ -2,7 +2,7 @@ public class Plant
 {
     public string Species { get; set; }
     public int LightNeeds { get; set; }
-    public double AskingPrice { get; set; }
+    public decimal AskingPrice { get; set; }
     public string City { get; set; }
     public string ZIP { get; set; }
     public bool IsSold { get; set; }
